pyproject.toml
[build-system]
requires = ["hatchling", "pynacl"]
build-backend = "hatchling.build"

[project]
name = "formsg-sdk"
version = "0.1.3"
authors = [
  { name="Shion Lee", email="shionlee92@gmail.com" },
]

dependencies = [
  "pynacl",
  "requests",
]

description = "Python SDK for FormSG"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
